Fix query from AppSync with an empty filter object
Build the query/mutation's selection set using the `graphql` utilities,
instead of building manually. This will avoid any edge cases with the
provided arguments. The schema is assumed to be valid: validation should
be taken care of ahead of time in `schemaModelValidator` and we also
currently do not define the built-in directives (e.g. `@id`,
`@relationship`) required to make it a valid schema.

- break out lookup functions for type/field definitions
- update `outputReference` resolvers
- add unit tests
- fix query result tests that were sensitive to object key order (use
  `toEqual` instead of strict JSON string equality)
- also check output file size of `.cjs` resolver in test `Case01.02`
- fix running test `Case01.04` (use `.cjs` extension)