Handle scalars better in `inferGraphDatabaseDirectives`
Make sure that nullability doesn't affect the inference of relationship
directives. Non-null lists of types can still have a filter/options
arguments. Non-null built-in scalars (e.g. `String`) no longer cause a
relationship directive and edge to be inferred.

The filter argument on a field will now refer to the scalar instead of
an input type, if appropriate.